Add `Scalar` and `Real` traits for basic number abstraction
This makes it a lot easier to write more functions as I'm not required
anymore to think about the exact trait bounds required. The latter is
also a problem because often, you can implement stuff in different ways
which would require different trait bounds. So the trait bounds kind of
expose implementation details.

In the end, this library's use case is interactive 3D applications. It
is not a general linear algebra library at all. So these restrictions
make my life a lot easier AND make the docs easier to read.