Add `{Vector, Point}::{truncate, extend}` by switching to nightly rustc
Unfortunately, the `N + 1` stuff is not stable yet. It is even marked
as incomplete feature and warns by default. One way I noticed this is
that type annotations for `T` in the doc examples are necessary.

But I think this stuff will be stabilized eventually and we can already
use it experimentally.