Make some more `[Hc]Matrix` method take `&self` instead of `self`
Ideally, I should perform some tests regarding compile time and
performance in debug and release mode. I'm not sure how bad it is to
copy these matrices around all the time. If it is bad, one should impl
all `ops` traits for references and treat the type like a non-`Copy`
type basically. But it is really inconvenient for users and clutters
the code with noisy `&` symbols.

So for now I just changed the places where it does not negatively affect
user code (e.g. because of auto ref of `self`).